// components/visit/visit.enum.ts
export enum EVisitSteps {
  PERSONAL_INFO = 'personalInfo',
  EXAMINATION = 'examination',
  TESTS = 'tests',
  NEXT_ACTION = 'nextAction',
}

export enum ESupportedTests {
  TEST_A = 'testA',
  TEST_B = 'testB',
  TEST_C = 'testC',
}

export enum EFormModes {
  CREATE = 'create',
  EDIT = 'edit',
}