// dto/createExamination.dto.ts
export interface ICreateExaminationDto {
  zoba: string;
  subZoba: string;
  address: string;
  bloodPressureDiastolic: string;
  bloodPressureSystolic: string;
  bloodSugar: string;
  heartRate: string;
  latitude: string;
  localDistrict: string;
  longitude: string;
  oxygenSaturation: string;
  respiratoryRate: string;
  temperature: string;
  hasFever: boolean;
  hasHeadache: boolean;
  hasDizziness: boolean;
  hasNausea: boolean;
  hasFatigue: boolean;
  hasWeightLoss: boolean;
  hasSweating: boolean;
  hasCough: boolean;
  hasShortnessOfBreath: boolean;
  hasSoreThroat: boolean;
  hasChestPain: boolean;
  hasVomiting: boolean;
  hasDiarrhea: boolean;
  hasStomachPain: boolean;
  hasConstipation: boolean;
  hasAppetiteLoss: boolean;
  hasMusclePain: boolean;
  hasJointPain: boolean;
  hasBackPain: boolean;
  hasNeckPain: boolean;
  hasNumbness: boolean;
  hasSeizures: boolean;
  hasDifficultySpeaking: boolean;
  hasRash: boolean;
  hasItching: boolean;
  hasBruising: boolean;
  hasPainfulUrination: boolean;
  hasFrequentUrination: boolean;
  hasBloodInUrine: boolean;
  hasEarPain: boolean;
  hasHearingLoss: boolean;
  hasNasalCongestion: boolean;
  hasRunnyNose: boolean;
  hasSneezing: boolean;
  hasEyePain: boolean;
  hasRedEye: boolean;
  hasBlurredVision: boolean;
  hasVisionLoss: boolean;
}